Refactored the timeseries handling to be more general and less transcription-specific. (#816)
* Pseudospectral and SolveIVP transcriptions have been generalized.  Hangup with the integration comp in ExplicitShooting...it needs an add_timeseries_output_configure that in turn reallocates the storage arrays.

* Some timeseries cleanup.  Parameters are now added to the timeseries outputs.
Colliding glob patterns will now result in a RuntimeError being raised during configure rather than a warning.

* fixed the transcirption source ODE path for SolveIVP

* fixed timeseries nodes, removed all non-default timeseries from SolveIVP

* ode-based timeseries outputs are now correctly added to the interleave comp for gauss-lobatto

* removed use of `dynamic` keywork for parameter when configuring timeseries outputs

* more work on SolveIVPTimeseries

* Radau timeseries configuration done.  Changed dymos.no_timeseries tag back to dymos.static_output.

* ExplicitShooting working, cleanup

* get_source_metadata now returns a single dict.  Some more cleanup

* Added output_name back as an argument to _get_timeseries_var_source...its needed by the GaussLobatto version.

* SolveIVPTimeseriesOutputComp correctly does conversion factors and allows for inputs to feed multiple outputs again.

* fix for unspecified state rate units

* Moved addition of state and state rates to the timeseries from the setup stack to configure.

* fixed source issue for the polynomial rates in the timeseries

* Fixed missing initial value data in parameter comp

* fixed handling of secondary timeseries

* cleanup

* SolveIVPTimeseriesComp now supports rate outputs, though they are necessarily not as accurate as the Pseudospectral one due to the equidistant time-spacing of nodes.
Added some checks of SolveIVP rate outputs to the min time climb tests.
Timeseries wildcard units are now handled correctly in the new system.

* Reworked the warning message when a timeseries output is not found in order to more closely match the old warning. Removed some of the logic that checks for timeseries outputs not in the ODE in the RKIntegrationComp since it was being triggered by non-ode-outputs like states and controls.

* SolveIVPTimeseriesOutputComp fix

* Fixed a bug in glob expansion of timeseries outputs.

* fixed a few side effects of recasing the get_source_metadata lookup error as a NameError instead of a RuntimeError.

* NameError -> ValueError to be a bit more pythonic

* new doc for tandem phases.  Fixed docstring of configure_states_introspection

* linting fix. added missing ipynb to docs

* Fixed bokeh install for docs

* more linting

* tandem phase doc cleanup

* tandem phase doc cleanup

* bokeh plot should display now in tandem brachistochrone doc

* fix for latest openmdao change

* edits from brets review

* one more change to set CI=0 for test_check_partials_no

* Fixed missing plots on commercial aircraft example docs.
Removed unnecessary _get_rate_source_path implelmentation for SolveIVP.

+++ dymos/utils/testing_utils.py
@@ -266,7 +266,7 @@ def assert_timeseries_near_equal(t_ref, x_ref, t_check, x_check, abs_tolerance=N
         raise ValueError("There is no overlapping time between the two time series")
 
     # Flatten the timeseries data arrays
-    num_elements = np.prod(shape_ref)
+    num_elements = np.prod(shape_ref, dtype=int)
     time_series_len = x_ref.shape[0]
     x_ref_data_flattened = np.reshape(x_ref, newshape=(time_series_len, num_elements))
     t_ref_unique, idxs_unique_ref = np.unique(t_ref.ravel(), return_index=True)
